Improved documentation
Added @throws in both method headers and javadocs where applicable for better documentation. Changed a few author tags to Jareth Batty as while the base sample was made by CS 1302 the rest of the project was made by Jareth Batty and thus it is more accurate to place him as the author. The sole exception to this is main which remains relativly intact from initial import thus the author remains the same.